requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
typer>=0.9.0